Update (#5)
* update product prices configuration

* update stock

* update transaction

* fix: handle supplier on null on transaction

* fix: supplier null on transaction

* add condition on stock form: disable the record has relation with transaction

* add default display unit for product

* report page

* update transaction

* access permission update

---------

+++ resources/views/productunit/datatables/action.blade.php
@@ -1,4 +1,8 @@
 <div class="row">
-    <a href="/unit-produk/edit/{{ $id }}"  class="btn btn-info btn-sm mr-1"><i class="fas fa-fw fa-pen"></i></a>
-    <a href="/unit-produk/delete/{{ $id }}" class="btn btn-danger btn-sm" data-confirm-delete="true"><i class="fas fa-fw fa-trash"></i></a>
+    @can('master-menu product-unit update')
+        <a href="/unit-produk/edit/{{ $id }}"  class="btn btn-info btn-sm mr-1"><i class="fas fa-fw fa-pen"></i></a>
+    @endif
+    @can('master-menu product-unit delete')
+        <a href="/unit-produk/delete/{{ $id }}" class="btn btn-danger btn-sm" data-confirm-delete="true"><i class="fas fa-fw fa-trash"></i></a>
+    @endif
 </div>
